requirements.txt
matplotlib
numpy
PyQt6
pyserial
